Implement threaded version of the incrementer
Neither of these "work" in the sense that they don't mess up the counter
like we wanted. I think the problem is that they do a _single_ increment
in each thread, which means that we never interrupt a thread in progress.
I think I need to have each thread do a bunch of increments, and then
they'll hopefully mess each other up.